Test completados :)

+++ DS-23-07-B2/src/e2/CompareSize.java
@@ -7,6 +7,6 @@ public class CompareSize implements Comparator<Vivienda> {
 
     @Override
     public int compare(Vivienda o1, Vivienda o2) {
-        return Integer.compare(o1.getSize(),o2.getSize());
+        return Float.compare(o1.getSize(),o2.getSize());
     }
 }
